+++ doc/code/targets/8_http_target.py
@@ -115,7 +115,7 @@ with RedTeamingOrchestrator(
     verbose=True,
     prompt_converters=[SearchReplaceConverter(old_value=r"(?! )\s", new_value="")],
 ) as red_teaming_orchestrator:
-    score = await red_teaming_orchestrator.apply_attack_strategy_until_completion_async(max_turns=3)  # type: ignore
+    score = await red_teaming_orchestrator.apply_attack_async(max_turns=3)  # type: ignore
     await red_teaming_orchestrator.print_conversation()  # type: ignore
 
 # %% [markdown]
