+++ pyrit/orchestrator/__init__.py
@@ -2,13 +2,14 @@
 # Licensed under the MIT license.
 
 from pyrit.orchestrator.orchestrator_class import Orchestrator
+from pyrit.orchestrator.multi_turn.multi_turn_orchestrator import MultiTurnOrchestrator
 
 from pyrit.orchestrator.crescendo_orchestrator import CrescendoOrchestrator
 from pyrit.orchestrator.flip_attack_orchestrator import FlipAttackOrchestrator
 from pyrit.orchestrator.fuzzer_orchestrator import FuzzerOrchestrator
 from pyrit.orchestrator.pair_orchestrator import PAIROrchestrator
 from pyrit.orchestrator.prompt_sending_orchestrator import PromptSendingOrchestrator
-from pyrit.orchestrator.red_teaming_orchestrator import RedTeamingOrchestrator
+from pyrit.orchestrator.multi_turn.red_teaming_orchestrator import RedTeamingOrchestrator
 from pyrit.orchestrator.scoring_orchestrator import ScoringOrchestrator
 from pyrit.orchestrator.skeleton_key_orchestrator import SkeletonKeyOrchestrator
 from pyrit.orchestrator.tree_of_attacks_with_pruning_orchestrator import TreeOfAttacksWithPruningOrchestrator
@@ -22,6 +23,7 @@ __all__ = [
     "CrescendoOrchestrator",
     "FlipAttackOrchestrator",
     "FuzzerOrchestrator",
+    "MultiTurnOrchestrator",
     "Orchestrator",
     "PAIROrchestrator",
     "PromptSendingOrchestrator",
